fix(fortify): cast sprintf argument to match %lu format specifier
count * FORTIFY_OVERHEAD promotes to size_t (64-bit on Win64) since
FORTIFY_OVERHEAD is sizeof-based, but the format string expects the
32-bit unsigned long the other arguments actually are. Cast it down
to match, fixing MSVC's format-string/argument-type mismatch warning.

+++ libraries/fortify/fortify.c
@@ -970,7 +970,7 @@ Fortify_LeaveScope(const char *file, unsigned long line)
   if (count)
   {
     sprintf(st_Buffer, "%10s %8lu bytes in %lu blocks with %lu bytes overhead\n",
-            "total", size, count, count * FORTIFY_OVERHEAD);
+            "total", size, count, (unsigned long) (count * FORTIFY_OVERHEAD));
     st_Output(st_Buffer);
   }
 
@@ -1029,7 +1029,7 @@ Fortify_ListAllMemory(const char *file, unsigned long line)
     }
 
     sprintf(st_Buffer, "%10s %8lu bytes in %lu blocks and %lu bytes overhead\n",
-            "total", size, count, count * FORTIFY_OVERHEAD);
+            "total", size, count, (unsigned long) (count * FORTIFY_OVERHEAD));
     st_Output(st_Buffer);
   }
